feat: support for user-submitted files
Previously, user-uploaded files were stored in the /openedx/xqueue/openedx
folder (considered by django to be a media subfolder). This folder was not
being mounted on the host, thus causing files to be lost on container restart.
Also, files were simply not reported by the `tutor submissions show` utility.

We now bind-mount the media folder from the host. Media assets are served by
uwsgi, which replaced gunicorn.

When object storage is in use, we will have to point xqueue to the remote
storage solution. This means that we need to add patches to the tutor-minio
plugin.

Close #2.